ref(iOS-Swift): extract sdk config out of app delegate (#4677)
