+++ Samples/iOS-Swift/iOS13-Swift/iOS13-Swift-Bridging-Header.h
@@ -1,4 +1,5 @@
 #import "SentryBenchmarking.h"
 #import "SentryUIApplication.h"
 #import <Sentry/PrivateSentrySDKOnly.h>
+#import <Sentry/SentryOptions+Private.h>
 #import <Sentry/SentrySDK+Private.h>
